#218 Inputwrapper is split to new files and new folder is made
