+++ src/components/accordions/Accordions.tsx
@@ -1,0 +1,111 @@
+import { Box, Button, Typography } from "@material-ui/core";
+import React, { FC, useEffect, useState } from "react";
+import AccordionComponent from "./AccordionComponent";
+import {
+  getListOfAttributes,
+  copyAttributeFromTemplateToApplication,
+} from "./copyAttribute";
+import AddCircleOutlineOutlinedIcon from "@material-ui/icons/AddCircleOutlineOutlined";
+import { useDocRef } from "../inputFields/saveInputFields";
+import { useRecoilValue } from "recoil";
+import { choosenApplicationState } from "../../stateManagement/choosenApplication";
+
+const Accordions: FC<AccordionsProps> = ({
+  title,
+  mainDesc,
+  chapterName,
+  attributeName,
+}) => {
+  const docRef = useDocRef();
+  const chosenApplication = useRecoilValue(choosenApplicationState);
+  const [newFields, setNewFields] = useState<any>([]);
+  const [isAccordionDeleted, setIsAccordionDeleted] = useState<boolean>(false);
+
+  useEffect(() => {
+    renderAccordions();
+    setIsAccordionDeleted(false);
+  }, [isAccordionDeleted]);
+
+  const accordionDeleted = (isDeleted: boolean) => {
+    setIsAccordionDeleted(isDeleted);
+  };
+
+  const renderAccordions = async () => {
+    let accordions: any = [];
+    await getListOfAttributes(docRef, attributeName, chapterName).then(
+      (attributeObjectList) => {
+        attributeObjectList.forEach((attribute) => {
+          Object.entries(attribute).forEach(([key, value]) => {
+            accordions.push(
+              <AccordionComponent
+                name={key}
+                inputFields={value.input_fields}
+                priority={value.priority}
+                title={title}
+                mainDesc={mainDesc}
+                chapterName={chapterName}
+                onAccordionDelete={accordionDeleted}
+              ></AccordionComponent>
+            );
+          });
+        });
+        accordions.sort((a: any, b: any) => a.key - b.key);
+        setNewFields(accordions);
+      }
+    );
+  };
+
+  const copyField = async (
+    docRef: any,
+    attributeName: string,
+    chapterName: string
+  ) => {
+    let attributeList = await copyAttributeFromTemplateToApplication(
+      chosenApplication,
+      docRef,
+      attributeName,
+      chapterName
+    );
+    var accordion;
+
+    Object.entries(attributeList[0]).forEach(([key, value]) => {
+      accordion = (
+        <AccordionComponent
+          name={key}
+          inputFields={value.input_fields}
+          priority={value.priority}
+          title={title}
+          mainDesc={mainDesc}
+          chapterName={chapterName}
+          onAccordionDelete={accordionDeleted}
+        ></AccordionComponent>
+      );
+    });
+
+    var accordions = [...newFields];
+    accordions.push(accordion);
+    setNewFields(accordions);
+  };
+  return (
+    <div>
+      <div>{newFields}</div>
+      <Box pl={1} mt={2}>
+        <Typography>
+          Husk å klikk på lagre om du har gjort noen endringer i en{" "}
+          {title.toLowerCase()}, før du sletter en annen {title.toLowerCase()}
+        </Typography>
+      </Box>
+      <Box mt={3} mb={1}>
+        <Button
+          onClick={() => copyField(docRef, attributeName, chapterName)}
+          variant="outlined"
+          startIcon={<AddCircleOutlineOutlinedIcon />}
+        >
+          Legg til {title}
+        </Button>
+      </Box>
+    </div>
+  );
+};
+
+export default Accordions;
