+++ src/components/inputFields/Date.tsx
@@ -1,5 +1,5 @@
 import React, { FC, useEffect, useRef, useState } from "react";
-import { useStyles } from "./inputStyles";
+import { useStyles } from "../../style/inputStyles";
 import { Typography, TextField, Box, InputLabel } from "@material-ui/core";
 import { useRecoilState } from "recoil";
 import { inputFieldObjectState } from "../../stateManagement/attributesState";
