+++ src/components/inputFields/getInputFieldComponent.tsx
@@ -1,0 +1,59 @@
+import React, { FC, useState } from "react";
+import Date from "./Date";
+import FileUpload from "./FileUpload";
+import LongText from "./LongText";
+import Number from "./Number";
+import RadioButton from "./RadioButton";
+import ShortText from "./ShortText";
+import Time from "./Time";
+
+export const componentList = [
+  { type: "short text", ComponentName: ShortText },
+  { type: "long text", ComponentName: LongText },
+  { type: "radio button", ComponentName: RadioButton },
+  { type: "date", ComponentName: Date },
+  { type: "time", ComponentName: Time },
+  { type: "file", ComponentName: FileUpload },
+  { type: "number", ComponentName: Number },
+];
+
+// Default component if nothing is selected
+const defaultComponent = () => {
+  return <div></div>;
+};
+
+const getComponentToBeRendered = (type: string) => {
+  let ComponentName: React.FC<{
+    desc: string;
+    id: string;
+    chapterName: string;
+  }>;
+  ComponentName = defaultComponent;
+
+  componentList.forEach((component) => {
+    if (component.type === type) {
+      ComponentName = component.ComponentName;
+    }
+  });
+
+  return ComponentName;
+};
+
+export const generateComponents = (
+  inputFields: Array<InputField>,
+  chapterName: string
+) => {
+  const components: any = [];
+  inputFields.map((inputField: any, i) => {
+    const Component = getComponentToBeRendered(inputField.type);
+    components.push(
+      <Component
+        key={i + "" + inputField.id}
+        desc={inputField.desc}
+        id={inputField.id}
+        chapterName={chapterName}
+      ></Component>
+    );
+  });
+  return components;
+};
