+++ src/react-app-env.d.ts
@@ -51,6 +51,14 @@ type InputWrapperProps = {
   priority: number;
 };
 
+type AccordionsProps = {
+  title: string;
+  mainDesc: string;
+  key: string;
+  chapterName: string;
+  attributeName: string;
+};
+
 type Attribute = {
   title: string;
   mainDesc: string;
@@ -85,7 +93,7 @@ type ErrorTypes = {
   text: ErrorText;
 };
 
-type InfoLongTextProps = {
+type HomeAccordionProps = {
   desc: string;
   priority: number;
   title: string;
@@ -96,4 +104,14 @@ type AppCardProps = {
   collectionName: string;
   to: string;
   onChange: (isUpdate: boolean) => void;
+};
+
+type AccordionProps = {
+  name: string;
+  inputFields: Array<InputField>;
+  priority: number;
+  title: string;
+  mainDesc: string;
+  chapterName: string;
+  onAccordionDelete: (isDeleted: boolean) => void;
 };
